Use jicoco for health checks.
