+++ pom.xml
@@ -265,6 +265,12 @@
       <version>${jersey.version}</version>
       <scope>test</scope>
     </dependency>
+      <dependency>
+          <groupId>org.jitsi</groupId>
+          <artifactId>jicoco-kotlin</artifactId>
+          <version>1.1-SNAPSHOT</version>
+          <scope>compile</scope>
+      </dependency>
   </dependencies>
 
   <build>
