+++ src/main/java/org/jitsi/jicofo/util/OsgiServiceBinder.java
@@ -1,6 +1,7 @@
 package org.jitsi.jicofo.util;
 
 import org.glassfish.hk2.utilities.binding.*;
+import org.jitsi.health.*;
 import org.jitsi.osgi.*;
 import org.osgi.framework.*;
 
@@ -19,5 +20,6 @@ public class OsgiServiceBinder extends AbstractBinder
         bind(new FocusManagerProvider(bundleContext)).to(FocusManagerProvider.class);
         bind(new JibriStatsProvider(bundleContext)).to(JibriStatsProvider.class);
         bind(new VersionServiceProvider(bundleContext)).to(VersionServiceProvider.class);
+        bind(new HealthCheckServiceProvider(bundleContext)).to(HealthCheckServiceProvider.class);
     }
 }
